use _Scene to wrap scene utilities
