+++ panel/scene-utils/index.html
@@ -1,0 +1,11 @@
+<script type="text/javascript" src="index.js"></script>
+
+<script type="text/javascript" src="engine-extends/playable/playable.js"></script>
+<script type="text/javascript" src="engine-extends/playable/ticker.js"></script>
+<script type="text/javascript" src="engine-extends/playable/time.js"></script>
+<script type="text/javascript" src="engine-extends/editor-engine.js"></script>
+<script type="text/javascript" src="engine-extends/debug-helper.js"></script>
+
+<script type="text/javascript" src="sandbox.js"></script>
+<script type="text/javascript" src="scene-undo.js"></script>
+<script type="text/javascript" src="engine-events.js"></script>
