+++ panel/scene-view.css
