+++ panel/scene-view.html
@@ -33,13 +33,5 @@
         Editor.require('packages://fire-gizmos/init.js');
     </script>
 
-    <script type="text/javascript" src="sandbox.js"></script>
-    <script type="text/javascript" src="scene-manager.js"></script>
-    <script type="text/javascript" src="playable/playable.js"></script>
-    <script type="text/javascript" src="playable/ticker.js"></script>
-    <script type="text/javascript" src="playable/time.js"></script>
-    <script type="text/javascript" src="editor-engine.js"></script>
-    <script type="text/javascript" src="debug-helper.js"></script>
-
     <script type="text/javascript" src="scene-view.js"></script>
 </dom-module>
