+++ panel/scene.html
@@ -1,5 +1,6 @@
 <link rel="import" href="packages://ui-kit/widget/index.html">
-<link rel="import" href="scene-view/scene-view.html">
+<link rel="import" href="scene-utils/index.html">
+<link rel="import" href="scene-view.html">
 
 <dom-module id="editor-scene">
     <link rel="import" type="css" href="scene.css">
